+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 Newest on top.
 
 ## [Unreleased]
+- "Party Finder" can now only be selected while the global switch is on.
 
 ## [v0.0.0.241] - 2026-06-22
 - "Hide nameplate" can now only be selected while the global switch is on.